Modifying behaviour of SubmitStorageRequest function so that the PC code does not invoke Guide.IsVisible.
Using the Guide (except for BeginShowStorageDeviceSelector/EndShowStorageDeviceSelector) requires a full XNA GS installation on the client computer.

In order to work around this, we assume on the PC it is set to true (and have a worker variable always set this way), whereas on the Xbox we will set the value to the actual value of Guide.IsVisible.